# docs/sad/03-design/openapi.yaml
openapi: 3.0.3
info:
  title: JPLearn Platform API
  version: 0.1.0
  description: 'Foundation v1. No grammar, flashcards, or L1 translation channels.


    Contract authority is this file (OpenAPI 3.0.3), not FastAPI-generated 3.1 JSON.

    Semantic normalized diff (ADR-003 D9): must match HTTP status, required/nullability,

    security, operationId, x-jplearn-fr, and error schema. Allowlist: 3.0↔3.1 nullable

    vs type union, generated component names, servers, info.


    Docs UI (`/docs`, `/redoc`, `/openapi.json`): off on staging and prod; local dev
    may

    enable. Contract tests read this file in git — do not depend on a public OpenAPI
    URL.

    '
servers:
- url: https://api.staging.jplearn.local
paths:
...
  /catalog:
    get:
      operationId: listCatalog
      tags:
      - Catalog
      x-jplearn-fr:
      - FR-CAT-002
      - FR-CAT-003
      - FR-CAT-004
      security:
      - bearerAuth: []
      parameters:
      - in: query
        name: ci_level
        schema:
          type: integer
          minimum: 0
          maximum: 4
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CatalogItemPublic'
  /catalog/search:
    get:
      summary: Search Catalog Scenes
      operationId: searchScenes
      tags:
      - Catalog
      x-jplearn-fr:
      - FR-SCH-001
      - FR-SCH-002
      - FR-NEG
      security:
      - bearerAuth: []
      parameters:
      - in: query
        name: q
        required: true
        description: Japanese search query
        schema:
          type: string
          minLength: 1
          maxLength: 100
          description: Japanese search query
      - in: query
        name: ci_level
        required: false
        description: Filter CI level
        schema:
          anyOf:
          - type: integer
            minimum: 0
            maximum: 4
          - type: 'null'
          description: Filter CI level
      - in: query
        name: cursor
        required: false
        description: Opaque pagination cursor
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Opaque pagination cursor
      - in: query
        name: limit
        required: false
        description: Page limit
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
          description: Page limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponsePublic'
        '400':
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
        '403':
          description: Scene search capability disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '409':
          description: Index generation mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
...
  /staff/catalog:
    get:
      operationId: listStaffCatalog
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CAT-005
      - FR-CMS-001
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - draft
            - level_qa
            - published
            - archived
            type: string
          - type: 'null'
          title: Status
      - name: ci_level
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 4
            minimum: 0
          - type: 'null'
          title: Ci Level
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogStaffList'
        '403':
          description: Not teacher or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    post:
      operationId: createCatalogItem
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CAT-001
      - FR-CAT-005
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogItemWrite'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemStaff'
        '403':
          description: Not teacher or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
  /staff/catalog/{id}:
    get:
      operationId: getStaffCatalogItem
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CAT-005
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemDetail'
        '403':
          description: Not teacher or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    patch:
      operationId: updateStaffCatalogItem
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CAT-005
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogItemPatch'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemStaff'
        '400':
          description: Item is not in draft status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '403':
          description: Not teacher or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
...
  /staff/catalog/{id}/submit-qa:
    post:
      operationId: submitLevelQa
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CMS-002
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: status=level_qa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemStaff'
  /staff/catalog/{id}/publish:
    post:
      operationId: publishCatalogItem
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CMS-002
      - FR-CMS-003
      - FR-CMS-004
      - FR-CAT-002
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: status=published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemStaff'
        '400':
          description: Not level_qa, or no media playback source (FR-CAT-002)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
  /staff/catalog/{id}/unpublish:
    post:
      operationId: unpublishCatalogItem
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CMS-002
      - FR-CAT-002
      description: Moves a published item back to draft so learners no longer see
        it.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: status=draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemStaff'
        '400':
          description: Item is not published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
...
  /staff/catalog/{id}/review:
    post:
      tags:
      - CMS
      summary: Review Catalog Item
      operationId: reviewCatalogItem
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogReviewBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemStaff'
        '400':
          description: Invalid review or workflow state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '403':
          description: Not teacher or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      x-jplearn-fr:
      - FR-CMS-002
      - NFR-SEC-002
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    signedQuery:
      type: apiKey
      in: query
      name: sig
      description: HMAC-SHA256 signature for media streaming (with exp timestamp)
  schemas:
...
    HttpError:
      type: object
      description: Standard JSON error body (400, 403, 404, 500) per ADR-005.
      required:
      - statusCode
      - message
      - error
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
          - type: string
          - type: array
            items:
              type: string
        error:
          type: string
    Http401Error:
      type: object
      description: Distinct 401 Unauthorized JSON error body per ADR-005.
      required:
      - statusCode
      - message
      properties:
        statusCode:
          type: integer
          enum:
          - 401
        message:
          type: string
...
    CatalogItemPublic:
      type: object
      description: Learner-facing. No L1 translation fields.
      required:
      - id
      - ci_level
      - duration_seconds
      - media_type
      - topic_id
      - visual_support
      properties:
        id:
          type: string
          format: uuid
        ci_level:
          type: integer
        duration_seconds:
          type: integer
        media_type:
          type: string
          enum:
          - video
          - audio
        topic_id:
          type: string
        visual_support:
          type: string
          enum:
          - high
          - medium
          - low
        playback_url:
          type: string
          format: uri
          nullable: true
        hls_url:
          type: string
          format: uri
          nullable: true
          description: NFR-PERF-002 HLS manifest when transcoded; clients prefer it
            and fall back to playback_url.
    CatalogItemWrite:
      type: object
      required:
      - topic_id
      - ci_level
      - duration_seconds
      - media_type
      - visual_support
      - title_internal
      properties:
        topic_id:
          type: string
        ci_level:
          type: integer
          description: v1 write has no min/max (Prisma Int). GET /catalog?ci_level=
            is 0–4.
        duration_seconds:
          type: integer
        media_type:
          type: string
          enum:
          - video
          - audio
        visual_support:
          type: string
          enum:
          - high
          - medium
          - low
        title_internal:
          type: string
    CatalogItemStaff:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        topic_id:
          type: string
          title: Topic Id
        ci_level:
          type: integer
          title: Ci Level
        duration_seconds:
          type: integer
          title: Duration Seconds
        media_type:
          type: string
          enum:
          - video
          - audio
          title: Media Type
        visual_support:
          type: string
          enum:
          - high
          - medium
          - low
          title: Visual Support
        title_internal:
          type: string
          title: Title Internal
        has_l1_translation:
          type: boolean
          const: false
          title: Has L1 Translation
        status:
          type: string
          enum:
          - draft
          - level_qa
          - published
          - archived
          title: Status
        revision:
          type: integer
          title: Revision
          default: 1
        qa_round:
          type: integer
          title: Qa Round
          default: 0
      additionalProperties: false
      type: object
      required:
      - id
      - topic_id
      - ci_level
      - duration_seconds
      - media_type
      - visual_support
      - title_internal
      - has_l1_translation
      - status
      title: CatalogItemStaff
    CatalogItemPatch:
      properties:
        revision:
          type: integer
          title: Revision
        topic_id:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Topic Id
        ci_level:
          anyOf:
          - type: integer
            maximum: 4.0
            minimum: 0.0
          - type: 'null'
          title: Ci Level
        duration_seconds:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Duration Seconds
        media_type:
          anyOf:
          - type: string
            enum:
            - video
            - audio
          - type: 'null'
          title: Media Type
        visual_support:
          anyOf:
          - type: string
            enum:
            - high
            - medium
            - low
          - type: 'null'
          title: Visual Support
        title_internal:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Title Internal
      additionalProperties: false
      type: object
      required:
      - revision
      title: CatalogItemPatch
    CatalogStaffList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemStaff'
...
    HighlightSpan:
      properties:
        start_offset:
          type: integer
          title: Start Offset
        end_offset:
          type: integer
          title: End Offset
      additionalProperties: false
      type: object
      required:
      - start_offset
      - end_offset
      title: HighlightSpan
    SearchResultItemPublic:
      properties:
        catalog_item_id:
          type: string
          title: Catalog Item Id
        content_version_id:
          type: string
          title: Content Version Id
        scene_id:
          type: string
          title: Scene Id
        scene_index:
          type: integer
          title: Scene Index
        start_time_seconds:
          type: integer
          title: Start Time Seconds
        end_time_seconds:
          type: integer
          title: End Time Seconds
        matched_text_ja:
          type: string
          title: Matched Text Ja
        highlight_spans:
          items:
            $ref: '#/components/schemas/HighlightSpan'
          type: array
          title: Highlight Spans
        match_kind:
          type: string
          enum:
          - exact_phrase
          - token_match
          title: Match Kind
        ci_level:
          type: integer
          title: Ci Level
        topic_id:
          type: string
          title: Topic Id
        title_jp:
          type: string
          title: Title Jp
      additionalProperties: false
      type: object
      required:
      - catalog_item_id
      - content_version_id
      - scene_id
      - scene_index
      - start_time_seconds
      - end_time_seconds
      - matched_text_ja
      - highlight_spans
      - match_kind
      - ci_level
      - topic_id
      - title_jp
      title: SearchResultItemPublic
    SearchResponsePublic:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SearchResultItemPublic'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        total_estimated:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Estimated
      additionalProperties: false
      type: object
      required:
      - items
      title: SearchResponsePublic
...
    CatalogReviewBody:
      properties:
        decision:
          type: string
          enum:
          - approve
          - reject
          title: Decision
        notes:
          type: string
          maxLength: 2000
          title: Notes
          default: ''
      additionalProperties: false
      type: object
      required:
      - decision
      title: CatalogReviewBody
...
    CatalogItemDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        topic_id:
          type: string
          title: Topic Id
        ci_level:
          type: integer
          title: Ci Level
        duration_seconds:
          type: integer
          title: Duration Seconds
        media_type:
          type: string
          enum:
          - video
          - audio
          title: Media Type
        visual_support:
          type: string
          enum:
          - high
          - medium
          - low
          title: Visual Support
        title_internal:
          type: string
          title: Title Internal
        has_l1_translation:
          type: boolean
          const: false
          title: Has L1 Translation
        status:
          type: string
          enum:
          - draft
          - level_qa
          - published
          - archived
          title: Status
        revision:
          type: integer
          title: Revision
          default: 1
        qa_round:
          type: integer
          title: Qa Round
        media:
          items:
            $ref: '#/components/schemas/MediaAssetStaff'
          type: array
          title: Media
        reviews:
          items:
            $ref: '#/components/schemas/CatalogReviewPublic'
          type: array
          title: Reviews
      additionalProperties: false
      type: object
      required:
      - id
      - topic_id
      - ci_level
      - duration_seconds
      - media_type
      - visual_support
      - title_internal
      - has_l1_translation
      - status
      - qa_round
      - media
      - reviews
      title: CatalogItemDetail